+++ 단건테스트/test.sh
@@ -44,7 +44,9 @@ fi
 ###############################################################
 # 30 Days over file Delete
 
-find $logdir/$svc -mtime +30 -print -exec rm -f {} \;
+if [ ! -z $logdir ];then
+	find $logdir/$svc -mtime +30 -print -exec rm -f {} \;
+fi
 
 ################### check Already  #############################
